ingestion/requirements.txt
fastapi
uvicorn[standard]
neo4j
python-dotenv
pydantic
scikit-learn
numpy
